+++ app/Http/Controllers/UserController.php
@@ -12,8 +12,10 @@ use App\Imports\UserCustomerImport;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
 use Maatwebsite\Excel\Facades\Excel;
+use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Request;
 use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Contracts\Encryption\DecryptException;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class UserController extends Controller
@@ -32,8 +34,8 @@ class UserController extends Controller
 
         if (request()->ajax()) {
             $query = User::query();
-
             return DataTables::of($query)
+
                 ->addColumn('roles', function ($item) {
                     if ($item->roles == 'USER') {
                         return '
@@ -67,24 +69,24 @@ class UserController extends Controller
                         ';
                     }
                 })
-                // ->addColumn('action', function ($item) {
-                //     return '
-                //     <div class="flex justify-start items-center space-x-3">
-                //     <a href="' . route('dashboard.user.edit', $item->id) . '" title="Edit"
-                //         class="flex flex-col shadow-sm items-center justify-center w-20 h-12 border border-yellow-500 bg-yellow-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline">
-                //         <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/edit.png') . '" alt="show" loading="lazy" width="20" />
-                //         <p class="mt-1 text-xs">Edit</p>
-                //     </a>
-                //     <form class="inline-block delete-form" action="' . route('dashboard.user.destroy', $item->id) . '" method="POST">
-                //         <button type="submit" title="Delete" class="inline-flex delete-form flex-col items-center justify-center w-20 h-12 bg-red-400 text-white rounded-md border border-red-500 transition duration-500 ease select-none hover:bg-red-500 focus:outline-none focus:shadow-outline">
-                //             <img class="w-6 h-6 rounded-full object-cover" src="' . asset('icon/delete.png') . '" alt="Delete" loading="lazy" width="20" />
-                //             <span class="mt-1 text-xs">Hapus</span>
-                //         </button>
-                //         ' . method_field('delete') . csrf_field() . '
-                //         </form>
-                // </div>
-                // ';
-                // })
+                ->addColumn('action', function ($item) {
+                    $encryptedId = Crypt::encrypt($item->id);
+                    return '
+                    <div class="flex justify-start items-center space-x-3.5">
+                    <a href="' . route('dashboard.user.edit', $encryptedId) . '" title="Edit"
+                        class="flex flex-col shadow-sm items-center justify-center w-20 h-12 border border-yellow-500 bg-yellow-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline">
+                        <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/edit.png') . '" alt="show" loading="lazy" width="20" />
+                        <p class="mt-1 text-xs">Edit</p>
+                    </a>
+                    <button type="button" title="Delete"
+                            class="flex flex-col delete-button shadow-sm items-center justify-center w-20 h-12 border border-red-500 bg-red-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-red-500 focus:outline-none focus:shadow-outline"
+                            data-id="' . $encryptedId . '">
+                            <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/delete.png') . '" alt="delete" loading="lazy" width="20" />
+                            <p class="mt-1 text-xs">Delete</p>
+                        </button>
+                </div>
+                ';
+                })
                 ->rawColumns(['roles', 'action'])
                 ->make();
         }
@@ -135,25 +137,24 @@ class UserController extends Controller
                         ';
                     }
                 })
-                // ->addColumn('action', function ($item) {
-                //     return '
-                //         <a title="Edit" class="inline-block border border-yellow-500 bg-yellow-400 text-white rounded-md px-2 py-1 m-1 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline"
-                //             href="' . route('dashboard.user.edit', $item->id) . '">
-                //             <div class="flex items-center">
-                //                 <span class="material-symbols-outlined inline-block mr-2">edit</span>
-                //                 <p class="inline-block">Edit</p>
-                //             </div>
-                //         </a>
-                //         <form class="inline-block" action="' . route('dashboard.user.destroy', $item->id) . '" method="POST">
-                //         <button title="Hapus" class="border border-red-500 bg-red-500 text-white rounded-md px-2 py-1 m-2 transition duration-500 ease select-none hover:bg-red-600 focus:outline-none focus:shadow-outline" >
-                //         <div class="flex items-center">
-                //                 <span class="material-symbols-outlined inline-block mr-2">delete</span>
-                //                 <p class="inline-block">Hapus</p>
-                //             </div>
-                //         </button>
-                //             ' . method_field('delete') . csrf_field() . '
-                //         </form>';
-                // })
+                ->addColumn('action', function ($item) {
+                    $encryptedId = Crypt::encrypt($item->id);
+                    return '
+                    <div class="flex justify-start items-center space-x-3.5">
+                    <a href="' . route('dashboard.user.edit', $encryptedId) . '" title="Edit"
+                        class="flex flex-col shadow-sm items-center justify-center w-20 h-12 border border-yellow-500 bg-yellow-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline">
+                        <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/edit.png') . '" alt="show" loading="lazy" width="20" />
+                        <p class="mt-1 text-xs">Edit</p>
+                    </a>
+                    <button type="button" title="Delete"
+                            class="flex flex-col delete-button shadow-sm items-center justify-center w-20 h-12 border border-red-500 bg-red-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-red-500 focus:outline-none focus:shadow-outline"
+                            data-id="' . $encryptedId . '">
+                            <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/delete.png') . '" alt="delete" loading="lazy" width="20" />
+                            <p class="mt-1 text-xs">Delete</p>
+                        </button>
+                </div>
+                ';
+                })
                 ->rawColumns(['roles', 'action'])
                 ->make();
         }
@@ -200,25 +201,25 @@ class UserController extends Controller
                         ';
                     }
                 })
-                // ->addColumn('action', function ($item) {
-                //     return '
-                //         <a title="Edit" class="inline-block border border-yellow-500 bg-yellow-400 text-white rounded-md px-2 py-1 m-1 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline"
-                //             href="' . route('dashboard.user.edit', $item->id) . '">
-                //             <div class="flex items-center">
-                //                 <span class="material-symbols-outlined inline-block mr-2">edit</span>
-                //                 <p class="inline-block">Edit</p>
-                //             </div>
-                //         </a>
-                //         <form class="inline-block" action="' . route('dashboard.user.destroy', $item->id) . '" method="POST">
-                //         <button title="Hapus" class="border border-red-500 bg-red-500 text-white rounded-md px-2 py-1 m-2 transition duration-500 ease select-none hover:bg-red-600 focus:outline-none focus:shadow-outline" >
-                //         <div class="flex items-center">
-                //                 <span class="material-symbols-outlined inline-block mr-2">delete</span>
-                //                 <p class="inline-block">Hapus</p>
-                //             </div>
-                //         </button>
-                //             ' . method_field('delete') . csrf_field() . '
-                //         </form>';
-                // })
+                ->addColumn('action', function ($item) {
+                    $encryptedId = Crypt::encrypt($item->id);
+                    return '
+                    <div class="flex justify-start items-center space-x-3.5">
+                    <button type="button" title="Edit"
+                            class="flex flex-col edit-button shadow-sm items-center justify-center w-20 h-12 border border-yellow-500 bg-yellow-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-yellow-500 focus:outline-none focus:shadow-outline"
+                            data-id="' . $encryptedId . '">
+                            <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/edit.png') . '" alt="delete" loading="lazy" width="20" />
+                            <p class="mt-1 text-xs">Edit</p>
+                        </button>
+                    <button type="button" title="Delete"
+                            class="flex flex-col delete-button shadow-sm items-center justify-center w-20 h-12 border border-red-500 bg-red-400 text-white rounded-md mx-2 my-2 transition duration-500 ease select-none hover:bg-red-500 focus:outline-none focus:shadow-outline"
+                            data-id="' . $encryptedId . '">
+                            <img class="object-cover w-6 h-6 rounded-full" src="' . asset('icon/delete.png') . '" alt="delete" loading="lazy" width="20" />
+                            <p class="mt-1 text-xs">Delete</p>
+                        </button>
+                </div>
+                ';
+                })
                 ->rawColumns(['roles', 'action'])
                 ->make();
         }
@@ -353,12 +354,26 @@ class UserController extends Controller
      * @param  \App\Models\User  $user
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function edit(User $user)
+    public function edit($encryptedId)
     {
-        return view('pages.dashboard.user.edit', [
-            'item' => $user
-        ]);
+        try {
+            $id = Crypt::decrypt($encryptedId);
+            $user = User::find($id);
+
+            if (!$user) {
+                // Lakukan penanganan jika pengguna tidak ditemukan
+                abort(404);
+            }
+
+            return view('pages.dashboard.user.edit', [
+                'item' => $user
+            ]);
+        } catch (\Illuminate\Contracts\Encryption\DecryptException $e) {
+            // Lakukan penanganan jika terjadi kesalahan dalam dekripsi
+            abort(400);
+        }
     }
+
 
     /**
      * Update the specified resource in storage.
@@ -382,13 +397,24 @@ class UserController extends Controller
      * @param  \App\Models\User  $user
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function destroy(User $user)
-    {
-        $user->delete();
 
-        return redirect()->route('dashboard.user.index')
-            // ->with('refresh', true)
-            // ->withSuccess('User berhasil dihapuskan!')
-        ;
+    public function destroy($encryptedId)
+    {
+        try {
+            $id = Crypt::decrypt($encryptedId);
+            $user = User::find($id);
+
+            if (!$user) {
+                // Lakukan penanganan jika user tidak ditemukan
+                abort(404);
+            }
+
+            $user->delete();
+
+            return redirect()->route('dashboard.user.index')->with('success', 'User telah dihapus.');
+        } catch (DecryptException $e) {
+            // Lakukan penanganan jika terjadi kesalahan dekripsi
+            abort(404);
+        }
     }
 }
